requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
qdrant-client==1.9.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0
tiktoken==0.5.2
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
